third_party/WebKit/LayoutTests/fast/workers/worker-onerror-09-expected.txt

This tests that unhandled exceptions in postMessage() are delivered to 'worker.onerror'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Page-level worker.onerror handler triggered:
PASS errorEvent.message is "Uncaught TypeError: Failed to execute 'postMessage' on 'DedicatedWorkerGlobalScope': The provided value cannot be converted to a sequence."
PASS stripURL(errorEvent.filename) is "[blob: URL]"
PASS errorEvent.lineno is 3
PASS errorEvent.colno is 9
PASS errorEvent.error is null
PASS successfullyParsed is true

TEST COMPLETE
